[Refactor] move `Result` unit test into new file: `result.test.ts`
